\name{addBarcodes}
\alias{addBarcodes}
\title{Generic Function to Add or Update Barcodes}
\description{
  A generic function to add or update barcodes in an object.
}
\usage{
  addBarcodes(object, barcodes)
}
\arguments{
  \item{object}{
    An object in which barcodes will be added or updated.
  }
  \item{barcodes}{
    A character vector of barcodes to be added or updated.
  }
}
\details{
  This is a generic function that will dispatch to the appropriate method depending on the class of the object.
}
\value{
  The result depends on the specific method for the class of the object.
}
