\name{cell_fate_DEG_calling}
\alias{cell_fate_DEG_calling}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
cell_fate_DEG_calling
}
\description{
This function calls differential expressed genes (DEG) between cells of specific cell types being members of asymmetric cell lineages (containing members of two or more cell types of interest) or members of symmetric cell lineages (contains members of only one cell type of interest). DEG calling is performed using Seurats FindMarkers function.
}
\usage{
cell_fate_DEG_calling(GEMLI_items, ident1, ident2, min.pct=0.05, logfc.threshold=0.1)
}
\arguments{
  \item{
  GEMLI_items}{GEMLI_items is a list of GEMLI inputs and outputs. To run 'cell_fate_DEG_calling' it should contain a 'gene_expression' as well as a 'cell_fate_analysis' element. The 'gene_expression' element is a quality controlled and normalised gene expression matrix where rownames are genes (features) and column names are cell IDs (cell barcodes). The 'cell_fate_analysis' element is a data frame with column 'cell.ID, 'clone.ID', 'cell.type' and 'cell.fate' generated by the 'extract_cell_fate_lineages' function.
  }
  \item{
  ident1}{'ident1' specifies the first 'cell.fate' of the GEMLI_items 'cell_fate_analysis' element to be used for DEG calling. It is a character vector which can encompass several cell fates. Cell fates contain the lineage type (sym or asym for symmetric or asymmetric respectively) and the cell type separated by an underscore. Cell fates can for example be 'sym_DCIS' or 'asym_inv_tumor'.
  }
  \item{
  ident2}{'ident2' specifies the second 'cell.fate' of the GEMLI_items 'cell_fate_analysis' element to be used for DEG calling. See ident1 for the format.
  }
  \item{
  min.pct}{'min.pct' is the min.pct parameter of Seurats FindMarker function. It is the minimum fraction of cells in either of the compared cell populations in which a gene should be expressed in order to be tested. the default value is 0.05. 
  }
  \item{
  logfc.threshold}{'logfc.threshold' is the logfc.threshold parameter of Seurats FindMarker function. It limits the DEG calling to genes which show, on average, at least x-fold differences (log-scale) between the two compared cell populations. 
  }
}
\value{
'cell_fate_DEG_calling' yields a data frame which is added to the 'GEMLI_items' under the name 'DEG'. The data frame is the output of Seurats FindMarker function and contains the column 'p_val', 'avg_log2FC', 'pct.1', 'pct2' and 'p_val_adj'.
}
\references{
A.S. Eisele*, M. Tarbier*, A.A. Dormann, V. Pelechano, D.M. Suter | "Gene-expression memory-based prediction of cell lineages from scRNA-seq datasets" | Nature Communications 15, 2744 (2024). https://doi.org/10.1038/s41467-024-47158-y
}
\author{
Almut Eisele and Marcel Tarbier
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
