\name{data_matrix}
\docType{data}
\alias{data_matrix}
\title{Example dataset for GEMLI}
\description{
    Example of gene expression table.
}
\usage{
data(GEMLI_example_data_matrix)
}
\format{
  A data frame with 32285 genes (rows) and 311 cells (columns).
}
\examples{
\dontrun{
  data(GEMLI_example_data_matrix)
  summary(data_matrix)
}
}
\keyword{datasets}
