\name{lineage_dict_bc}
\docType{data}
\alias{lineage_dict_bc}
\title{Example dataset for GEMLI}
\description{
    Named integer with 156 values and cell ids as names.
}
\usage{
data(GEMLI_example_barcode_information)
}
\format{
  A data frame with 32285 genes (rows) and 311 cells (columns).
}
\examples{
\dontrun{
  data(GEMLI_example_barcode_information)
  summary(lineage_dict_bc)
}
}
\keyword{datasets}
