\name{potential_markers}
\alias{potential_markers}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
potential_markers
}
\description{
Add description here !
}
\usage{
potential_markers(data_matrix)
}
\arguments{
  \item{
  data_matrix}{Explain argument !}
}
\value{
Explain output !
}
\references{
A.S. Eisele*, M. Tarbier*, A.A. Dormann, V. Pelechano, D.M. Suter | "Gene-expression memory-based prediction of cell lineages from scRNA-seq datasets" | Nature Communications 15, 2744 (2024). https://doi.org/10.1038/s41467-024-47158-y
}
\author{
Marcel Tarbier and Almut Eisele
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
}
